@@ -32,12 +32,13 @@
     "nodemon": "^3.0.3",
     "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.1",
+    "react-diff-viewer": "^3.1.1",
     "react-dom": "^18.2.0",
     "react-horizontal-scrolling-menu": "^4.1.0",
+    "react-json-view-compare": "^2.0.2",
     "react-router-dom": "^6.19.0",
     "react-scripts": "5.0.1",
     "seamless-scroll-polyfill": "^2.3.4",
-    "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -65,5 +66,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
   }
 }
